_cards: Define @generate-card mixin
``@generate-card`` helps generating cards by taking
their dimensions, background color, their name and
a boolean specifying whether or not the card as an image
on top.

Adds ``.card-list`` which ensures that the cards are
aligned horizontally.

Add ``.card-desc-small``, ``.card-desc-small-text`` and
``.card-desc-small-img`` which are the classes used to
style the cards in the what-we-do secion.